GH-1339 renamed controller
