+++ SafeUIKit/SafeAppUI/MainFlow/Menu/WalletConnect/WCCompletionPanelViewController.xib
@@ -8,9 +8,9 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="WCCompletedViewController" customModule="SafeAppUI" customModuleProvider="target">
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="WCCompletionPanelViewController" customModule="SafeAppUI" customModuleProvider="target">
             <connections>
-                <outlet property="hudView" destination="ugm-HP-FDM" id="7rp-9T-MIN"/>
+                <outlet property="panelContentView" destination="ugm-HP-FDM" id="7rp-9T-MIN"/>
                 <outlet property="textLabel" destination="ubx-Pd-NCa" id="dke-cY-WhX"/>
                 <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
             </connections>
